SAMZA-2583: Fix NPE in update SamzaHistogram Gauge values
Symptopm: NullPointerException when creating SamzaHistogram

Cause: SamzaHistogram gauges are instantiated via passed in metric
registry newGauge api. The new Gauge api in turn tries to instantiate
HistogramGauge. The newGauge api calls getValue on HistogramGauge
which in turn tries to update the Histogram Gauge values. This means
we are updating Histogram Gauge values in process of creating it.

Changes:As a fix, we add  null check in Gauge getValue to avoid NPE.

Tests: Added unit tests to check for the same.